refactor: change stucture of auth feature
