+++ src/app/app-layout/AppLayout.ts
@@ -1,6 +1,6 @@
 import { Component, Header, main } from "../../shared";
 import { Router } from "../router/router";
-import { AuthState } from "../../features/auth/AuthState";
+import { AuthState } from "../../features/auth/model/AuthState";
 
 import styles from "./AppLayout.module.css";
