--- src/features/auth/types.ts
@@ -1,4 +1,0 @@
-export interface UserCredentials {
-  firstName: string;
-  surname: string;
-}
