+++ src/pages/login/LoginPage.ts
@@ -1,8 +1,8 @@
-import { AuthState } from "../../features/auth/AuthState";
+import { AuthState } from "../../features/auth/model/AuthState";
 import { Router } from "../../app/router/router";
 
 import { Component, h2 } from "../../shared";
-import { LoginForm } from "../../features/auth/LoginForm";
+import { LoginForm } from "../../features/auth/ui/LoginForm";
 
 import styles from "./LoginPage.module.css";
